+++ apps/storybook/.storybook/preview.ts
@@ -1,5 +1,5 @@
 import type { Preview } from '@storybook/react-vite';
-// import '@repo/ui/global.css';
+import '@repo/ui/styles.css';
 
 const preview: Preview = {
   parameters: {
