+++ packages/ui/src/global.css
@@ -1,4 +1,6 @@
-@import 'tailwindcss';
+@tailwind base;
+@tailwind components;
+@tailwind utilities;
 @import 'tw-animate-css';
 
 :root {
