+++ turbo.json
@@ -30,6 +30,10 @@
     "build-storybook": {
       "dependsOn": ["^build-storybook"],
       "inputs": ["$TURBO_DEFAULT$", "storybook.config.*"]
+    },
+    "storybook": {
+      "dependsOn": ["^build"],
+      "cache": false
     }
   },
   "globalDependencies": ["package.json", "pnpm-workspace.yaml", "tsconfig.json"],
